Move generic string helper functions to dedicated file.
