+++ lib/application_business_rules/command_line_interpreter.hpp
@@ -1,28 +1,9 @@
-#include <cstring>
-#include <cwchar>
 #include <functional>
-#include <iomanip>
 #include <iostream>
 #include <iterator>
-#include <sstream>
 #include <stdexcept>
+#include <string_helpers.hpp>
 #include <tuple>
-#include <vector>
-
-template <typename CharType>
-int strcmp_g(const CharType *const lhs, const CharType *const rhs);
-
-template <>
-int strcmp_g(const char *const lhs, const char *const rhs)
-{
-    return std::strcmp(lhs, rhs);
-}
-
-template <>
-int strcmp_g(const wchar_t *const lhs, const wchar_t *const rhs)
-{
-    return std::wcscmp(lhs, rhs);
-}
 
 namespace command_line_interpreter
 {
@@ -56,27 +37,6 @@ struct Option
 
     T defaultValue;
 };
-
-// Tokenize a string using std::quoted
-template <typename CharType>
-std::vector<std::basic_string<CharType>> tokenizeQuoted(const std::basic_string<CharType> &input)
-{
-    std::basic_stringstream<CharType> iss(input);
-    std::vector<std::basic_string<CharType>> tokens;
-
-    std::basic_string<CharType> token;
-    while (!iss.eof())
-    {
-        iss >> std::quoted(token);
-        if (iss.fail())
-        {
-            throw std::runtime_error("failed to tokenize string: '" + input + "'");
-        }
-        tokens.push_back(token);
-    }
-
-    return tokens;
-}
 
 /**
  * Combines a command with a function.
